remove FOSRest on symfony 5 test runs as FOSRest does not support symfony 5 yet
